Added drawing support and changed graphics lib
This commit adds drawing support to the AVL tree and switches the graphics
engine to OpenGL. Drawing support functions by generating a vector of points and
then drawing lines to connect those points.
This commit also fixes a bug in the speech presentation of node items; values
are now properly presented and the system generates doubles instead of uint16s.

Signed-off-by: Ethin Probst <ethindp@protonmail.com>